\name{MATLAB}
\alias{matlab}
\title{MATLAB replacement functions}
\description{
  Wrapper functions and variables used to replicate MATLAB
function calls as best possible to simplify porting.
}
\details{
   They are no more complete than absolutely necessary and are quite
possibly broken for fringe cases.
}
\note{
   In certain cases, these may not correspond exactly with MATLAB API
as sometimes it just wasn't possible.
}
\author{P. Roebuck, \email{roebuck@mdanderson.org}}
\keyword{interface}

