\name{eye}
\alias{eye}
\title{MATLAB eye function}
\description{
  Create an identity matrix.
}
\usage{
eye(n, m = n)
}
\arguments{
  \item{n, m}{dimensions for the result}
}
\value{
   Returns \code{matrix} of order 1.
   Defaults to square if second dimension argument \code{m} not provided.
}
\examples{
eye(3)
}
\author{P. Roebuck, \email{roebuck@mdanderson.org}}
\keyword{interface}

